chore: move auth and auth config to server dir
