+++ src/components/Note/Notes.tsx
@@ -5,7 +5,7 @@ import SearchNote from './SearchNote';
 import SortDropdown from '../SortDropdown';
 import { getFilteredNotes, getFilter } from '@/utils/format-notes';
 import { getAllUserOrdinaryNotes } from '@/queries/note';
-import { auth } from '@/auth';
+import { auth } from '@/auth/auth';
 import { ReactNode } from 'react';
 
 export default async function Notes() {
