+++ src/middleware.ts
@@ -6,7 +6,7 @@ import {
   publicRoutes,
 } from './routes';
 import NextAuth from 'next-auth';
-import authConfig from './auth.config';
+import authConfig from '@/auth/auth.config';
 
 const { auth } = NextAuth(authConfig);
