+++ src/server/auth/auth.ts
@@ -1,6 +1,6 @@
 import NextAuth from 'next-auth';
 import { PrismaAdapter } from '@auth/prisma-adapter';
-import authConfig from '@/auth.config';
+import authConfig from '@/auth/auth.config';
 import { getUserById } from '@/queries/user';
 import { createPlaceholderNote } from '@/actions/note';
 import { db } from '@/server/db';
